+++ CHANGELOG.md
@@ -7,12 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.0] - 2023-10-14
+
 ### Added
 
 - [Project history](src/docs/common/history.md) section
 - [Doc Format Summary](src/docs/common/doc_format_summary.md) entry point section
+- [Doc Format and Apache FreeMarker](src/docs/common/doc_format_freemarker.md) guidelines
 - [From XML to JSON or YAML source](src/docs/common/doc_from_xml_to_json_or_yaml.md) section
 - [Venus Guide E001 : Hello World PDF JSON Source](fj-doc-guides-E001-hello-world-pdf-json-source/README.md)
 - Started guidelines section
 - [Venus Guide A001 : Hello World PDF](fj-doc-guides-A001-hello-world-pdf/README.md)
+- [Venus Guide A002 : Full document](fj-doc-guides-A002-full-document/README.md)
+- [Venus Guide A003 : Apache FreeMarker Full document](fj-doc-guides-A003-full-document-freemarker/README.md)
+- [Venus Guide A004 : Apache FreeMarker Hello World](fj-doc-guides-A004-freemarker-hello-world/README.md)
 - badges to README.md
+- Added [FreeMarker](fj-doc-guides-A003-full-document-freemarker/README.md)